+++ customer_frontend/src/components/handleProducts/ViewProducts.js
@@ -144,6 +144,7 @@ const useToolbarStyles = makeStyles((theme) => ({
         },
   title: {
     flex: '1 1 100%',
+    fontFamily:'Arial Black'
   },
 }));
 
@@ -355,8 +356,18 @@ useEffect(() => {
 
   const handleDelete = (_id) => {
 
-  
-    console.log(selected.toString(_id))
+  //console.log(selected.toString(_id))
+  Swal.fire({
+    title: 'Are you sure ?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes!'
+  })
+  .then((result) => {
+    if (result.isConfirmed) {
     axios.delete(
       `http://localhost:5000/products/` + selected.toString(_id)
     )
@@ -379,8 +390,15 @@ useEffect(() => {
             title: 'Oops...',
             text: 'Something went wrong!',
           })
+
     }
     })
+    }else{
+      //getProductData();
+      setSelected([]);
+    }
+  })
+
   
   }
